backend/gemini-artistic-api/requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Google Cloud
google-cloud-storage>=2.14.0
google-cloud-firestore>=2.14.0
google-generativeai>=0.8.0

# Configuration and utilities
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1

# Image processing
Pillow>=10.2.0

# HTTP client
httpx>=0.26.0
aiofiles>=24.1.0
